# client: add the fuse parameter max_write