This parameter has the following 2 advantages:
1.  FUSE(version > 2.8) the default single IO write size is 128k (controlled by max_write), if I use bs=4M in 
    the FIO tool test,you will find that 4*1024k/128=32 is needed, ie 32 IO operations are needed . If I 
    adjust max_write to 4M,only one operation is needed, which greatly improves the write performance 
    of cephfs during fuse mount.Of course, the above implementation requires libfuse and kernel fuse to 
    support.
2. In addition, we can also limit the single IO write size by setting max_write to less than 128K.

Signed-off-by: huanwen ren <ren.huanwen@zte.com.cn>