@renhwztetecs 2018-03-13
Before:
ceph-fuse -k /etc/ceph/ceph.client.admin.keyring -m 10.118.202.181:6789 /mycephfs

> fio-filename=/mycephfs/fio_4M-1 -direct=1 -iodepth 1 -thread -rw=write -ioengine=libaio -bs=4M -size=1g -numjobs=1 -runtime=300 -group_reporting -name=testceph  
cephfs: (g=0): rw=write, bs=4M-4M/4M-4M/4M-4M, ioengine=libaio, iodepth=4
...
fio-2.12
Starting 1 threads
cephfs: Laying out IO file(s) (1 file(s) / 512MB)
Jobs: 1 (f=1): [_(2),W(1),_(1)] [97.0% done] [0KB/32702KB/0KB /s] [0/7/0 iops] [eta 00m:18s]
cephfs: (groupid=0, jobs=1): err= 0: pid=19209: Tue Mar  6 17:23:16 2018
  write: io=2048.0MB, bw=3593.4KB/s, iops=0, runt=583671msec
    slat (msec): min=1188, max=9891, avg=4531.13, stdev=1449.52
    clat (usec): min=8, max=20555K, avg=13325163.34, stdev=3302525.98
     lat (msec): min=2716, max=25982, avg=17856.30, stdev=3656.44
    clat percentiles (msec):
     |  1.00th=[ 2769],  5.00th=[ 9110], 10.00th=[ 9503], 20.00th=[10683],
     | 30.00th=[11600], 40.00th=[12649], 50.00th=[13304], 60.00th=[13960],
     | 70.00th=[14746], 80.00th=[16057], 90.00th=[16712], 95.00th=[16712],
     | 99.00th=[16712], 99.50th=[16712], 99.90th=[16712], 99.95th=[16712],
     | 99.99th=[16712]
    bw (KB  /s): min= 8175, max= 8342, per=100.00%, avg=8213.09, stdev=23.94
    lat (usec) : 10=0.59%, 20=0.20%
    lat (msec) : >=2000=99.22%
  cpu          : usr=0.01%, sys=0.02%, ctx=17442, majf=0, minf=58034
  IO depths    : 1=0.8%, 2=1.6%, 4=97.7%, 8=0.0%, 16=0.0%, 32=0.0%, >=64=0.0%
     submit    : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     issued    : total=r=0/w=512/d=0, short=r=0/w=0/d=0, drop=r=0/w=0/d=0
     latency   : target=0, window=0, percentile=100.00%, depth=4
Run status group 0 (all jobs):
  WRITE: io=2048.0MB, **aggrb=3593KB/s**, minb=3593KB/s, maxb=3593KB/s, mint=583671msec, maxt=583671msec



After:
ceph-fuse -o big_writes -o max_write=4194304 -k /etc/ceph/ceph.client.admin.keyring -m 10.118.202.181:6789 /mycephfs

> fio-filename=/mycephfs/fio_4M-1 -direct=1 -iodepth 1 -thread -rw=write -ioengine=libaio -bs=4M -size=1g -numjobs=1 -runtime=300 -group_reporting -name=testceph  
testceph: (g=0): rw=write, bs=4M-4M/4M-4M/4M-4M, ioengine=libaio, iodepth=1
fio-2.12
Starting 1 thread
testceph: Laying out IO file(s) (1 file(s) / 1024MB)
Jobs: 1 (f=1): [W(1)] [100.0% done] [0KB/28672KB/0KB /s] [0/7/0 iops] [eta 00m:00s]
testceph: (groupid=0, jobs=1): err= 0: pid=2862: Mon Mar 12 22:03:57 2018
  write: io=1024.0MB, bw=37925KB/s, iops=9, runt= 27649msec
    slat (msec): min=70, max=299, avg=108.00, stdev=35.54
    clat (usec): min=0, max=9, avg= 1.52, stdev= 0.90
     lat (msec): min=70, max=299, avg=108.00, stdev=35.54
    clat percentiles (usec):
     |  1.00th=[    1],  5.00th=[    1], 10.00th=[    1], 20.00th=[    1],
     | 30.00th=[    1], 40.00th=[    1], 50.00th=[    1], 60.00th=[    1],
     | 70.00th=[    2], 80.00th=[    2], 90.00th=[    2], 95.00th=[    3],
     | 99.00th=[    3], 99.50th=[    8], 99.90th=[    9], 99.95th=[    9],
     | 99.99th=[    9]
    bw (KB  /s): min=16384, max=49152, per=100.00%, avg=37981.09, stdev=7286.95
    lat (usec) : 2=60.16%, 4=39.06%, 10=0.78%
  cpu          : usr=0.17%, sys=0.04%, ctx=515, majf=0, minf=4
  IO depths    : 1=100.0%, 2=0.0%, 4=0.0%, 8=0.0%, 16=0.0%, 32=0.0%, >=64=0.0%
     submit    : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     issued    : total=r=0/w=256/d=0, short=r=0/w=0/d=0, drop=r=0/w=0/d=0
     latency   : target=0, window=0, percentile=100.00%, depth=1
Run status group 0 (all jobs):
  WRITE: io=1024.0MB, aggrb=**37924KB/s**, minb=37924KB/s, maxb=37924KB/s, mint=27649msec, maxt=27649msec